refactor(solver): split mapped relation expansion helper
