+++ crates/tsz-solver/src/relations/subtype/rules/mod.rs
@@ -11,6 +11,7 @@
 //! - `objects`: Object property matching and index signatures
 //! - `functions`: Function/callable signature compatibility
 //! - `generics`: Type parameters, references, and applications
+//! - `mapped_expansion`: Concrete mapped-type relation expansion
 //! - `conditionals`: Conditional type checking
 
 pub mod conditionals;
@@ -19,6 +20,7 @@ pub mod generics;
 pub mod intrinsic_object;
 pub mod intrinsics;
 pub mod literals;
+pub mod mapped_expansion;
 pub mod mapped_key_constraints;
 pub mod objects;
 pub mod promise_like;
